Update unit tests & mock data
