--- src/applications/dependents/686c-674/tests/config/chapters/report-add-child/relationship.unit.spec.jsx
@@ -1,43 +1,0 @@
-import { render } from '@testing-library/react';
-import { Provider } from 'react-redux';
-import { expect } from 'chai';
-import React from 'react';
-import createCommonStore from '@department-of-veterans-affairs/platform-startup/store';
-import {
-  DefinitionTester,
-  getFormDOM,
-} from 'platform/testing/unit/schemaform-utils';
-import formConfig from '../../../../config/form';
-
-const defaultStore = createCommonStore();
-
-const formData = {
-  'view:selectable686Options': {
-    addChild: true,
-  },
-  childrenToAdd: [{}],
-};
-describe('686 add child relationship part one', () => {
-  const {
-    schema,
-    uiSchema,
-  } = formConfig.chapters.addChild.pages.addChildRelationshipPartOne;
-
-  it('should render', () => {
-    const form = render(
-      <Provider store={defaultStore}>
-        <DefinitionTester
-          schema={schema}
-          definitions={formConfig.defaultDefinitions}
-          uiSchema={uiSchema}
-          data={formData}
-          arrayPath="childrenToAdd"
-          pagePerItemIndex={0}
-        />
-      </Provider>,
-    );
-
-    const formDOM = getFormDOM(form);
-    expect(formDOM.querySelectorAll('va-radio-option').length).to.eq(2);
-  });
-});
