+++ src/applications/dependents/686c-674/tests/e2e/fixtures/686c-only-maximal.json
@@ -153,6 +153,7 @@
       },
       "biologicalParentSsn": "333444554",
       "biologicalParentDob": "2014-09-19",
+      "relationshipType": "STEPCHILD",
       "relationshipToChild": {
         "adopted": true,
         "stepchild": true
