+++ src/applications/dependents/686c-674/tests/e2e/fixtures/maximal.json
@@ -210,6 +210,7 @@
       },
       "biologicalParentSsn": "333444554",
       "biologicalParentDob": "2014-09-19",
+      "relationshipType": "STEPCHILD",
       "relationshipToChild": {
         "adopted": true,
         "stepchild": true
