+++ src/applications/dependents/686c-674/tests/e2e/fixtures/picklist.json
@@ -194,6 +194,7 @@
       },
       "biologicalParentSsn": "333444554",
       "biologicalParentDob": "2014-09-19",
+      "relationshipType": "STEPCHILD",
       "relationshipToChild": {
         "adopted": true,
         "stepchild": true
