+++ src/applications/dependents/686c-674/tests/e2e/fixtures/transformed-maximal.json
@@ -157,6 +157,7 @@
       "biologicalParentName": { "first": "John", "last": "Doe" },
       "biologicalParentSsn": "333444554",
       "biologicalParentDob": "2014-09-19",
+      "relationshipType": "STEPCHILD",
       "relationshipToChild": { "adopted": true, "stepchild": true },
       "isBiologicalChild": false,
       "birthLocation": {
